+++ demos/middleware/netlify.toml
@@ -3,12 +3,8 @@ command = "npm run build"
 publish = ".next"
 ignore = "if [ $CACHED_COMMIT_REF == $COMMIT_REF ]; then (exit 1); else git diff --quiet $CACHED_COMMIT_REF $COMMIT_REF ../..; fi;"
 
-[build.environment]
-NEXT_USE_NETLIFY_EDGE = "true"
-
-[[plugins]]
-package = "../plugin-wrapper/"
-
+# [[plugins]]
+# package = "../plugin-wrapper/"
 # This is a fake plugin, that makes it run npm install
 [[plugins]]
 package = "@netlify/plugin-local-install-core"
@@ -18,6 +14,7 @@ included_files = [
     "!node_modules/sharp/vendor/8.12.2/darwin-*/**/*",
     "!node_modules/sharp/build/Release/sharp-darwin-*"
 ]
+# Uncomment this if testing the built files rather than dev
 # [dev]
 # framework = "#static"
 # [[redirects]]
