+++ demos/middleware/tsconfig.json
@@ -17,7 +17,6 @@
   },
   "exclude": [
     "node_modules",
-    "netlify/edge-functions/**/*",
     "../../src/templates/edge/*"
   ],
   "include": [
